+++ hand_items/hand_item.lua
@@ -14,8 +14,8 @@ function HandItem:newKnife()
         angle = 0,
         sprite = Assets.quads.knife[1],
         static = true,
-        walkSpeed = 130, -- TUNE: knife = fastest
-        damage = 60 -- TUNE: used when melee lands (step 8)
+        walkSpeed = TUNE.knife.walkSpeed,
+        damage = TUNE.knife.damage -- used when melee lands (step 8)
     }
 
     setmetatable(obj, HandItem)
@@ -31,8 +31,8 @@ function HandItem:newGrenade(type)
         sprite = Assets.quads.grenade[1],
         static = true,
         type = type or 'he',
-        walkSpeed = 120, -- TUNE
-        damage = 120 -- TUNE: full damage anywhere inside blast radius (step 9)
+        walkSpeed = TUNE.grenade.walkSpeed,
+        damage = TUNE.grenade.damage -- full damage anywhere inside blast radius (step 9)
     }
 
     setmetatable(obj, HandItem)